+++ src/object/any.js
@@ -1,4 +1,5 @@
 import _curry2 from '../_internals/_curry2.js'
+import keys from './keys.js'
 
 /**
  * @name any
@@ -23,7 +24,7 @@ import _curry2 from '../_internals/_curry2.js'
  * run({ a: 'xxx', b: 'bar' }) // => false
  */
 const any = (schema, obj) =>
-  Object.keys(schema).some(key =>
+  keys(schema).some(key =>
     schema[key](obj[key]))
 
 export default _curry2(any)