+++ tests/function/gt.js
@@ -3,11 +3,13 @@ import test from 'tape'
 
 test('gt -- Determines if the numbers are gt', t => {
   t.true(gt(1, 2))
+  t.false(gt(7.0, 7.0))
   t.end()
 })
 
 test('gt -- Determines if the letters are gt', t => {
   t.true(gt('a', 'b'))
+  t.true(gt('abc', 'abcd'))
   t.end()
 })